pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloudpop"
version = "0.1.0"
description = "Multi-provider cloud media bridge: convert cloud storage videos to STRM files for Plex/Skybox VR playback"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "click>=8.1.7",
    "pydantic-settings>=2.5.0",
    "pyyaml>=6.0.2",
    "cachetools>=5.5.0",
    "anyio>=4.6.0",
    "p115cipher>=0.0.4",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.21.1",
    "ruff>=0.7.0",
]

[project.scripts]
cloudpop = "cloudpop.cli.commands:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["cloudpop*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
